+++ src/main/java/test.java
@@ -7,20 +7,4 @@ import org.junit.Test;
 
 public class test {
 
-    private PropertyReader propertyReader;
-    private Driver driver;
-
-    @Before
-    public void startDriver(){
-        propertyReader = new PropertyReader();
-        Driver.createDriver(DriverType.IE);
-        Driver.getDriver().manage().window().maximize();
-    }
-
-    @Test
-    public void x(){
-        Driver.getDriver().get("https://www.google.com/");
-
-    }
-
 }
